ai_engine/requirements.txt
accelerate
aiofiles
annotated-doc
annotated-types
antlr4-python3-runtime
anyio
attrs
brotli
certifi
charset-normalizer
click
colorama
coloredlogs
diffusers
einops
exceptiongroup
fastapi
ffmpy
filelock
flatbuffers
fsspec
glcontext
gradio
gradio_client
groovy
h11
hf-xet
httpcore
httpx
huggingface_hub
humanfriendly
idna
ImageIO
importlib_metadata
Jinja2
jsonschema
jsonschema-specifications
lazy_loader
llvmlite
markdown-it-py
MarkupSafe
mdurl
moderngl
mpmath
networkx
numba
numpy
omegaconf
onnxruntime
opencv-python
opencv-python-headless
orjson
packaging
pandas
pillow
platformdirs
pooch
protobuf
psutil
pydantic
pydantic_core
pydub
pyglet
Pygments
PyMatting
PyMCubes
pyreadline3
python-dateutil
python-multipart
pytz
PyYAML
referencing
regex
rembg
requests
rich
rpds-py
safehttpx
safetensors
scikit-image
scipy
semantic-version
shellingham
six
starlette
sympy
tifffile
tokenizers
tomlkit
torch
torchmcubes @ git+https://github.com/tatsy/torchmcubes.git@3381600ddc3d2e4d74222f8495866be5fafbace4
torchvision
tqdm
transformers
trimesh
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
xatlas
zipp
